requirements.txt
# Generate exact requirements
pip freeze > requirements-lock.txt

# Or create clean requirements.txt
cat > requirements.txt << 'EOF'
# Document Processing
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
lxml>=5.0.0

# Text Processing
tiktoken>=0.5.0

# ML/Embeddings
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
huggingface_hub==0.16.4

# CLI
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Async
aiofiles>=23.0.0

# PDF Generation (for tests)
reportlab>=4.0.0
EOF